[in_app_purchase] Iap refactor (#1381)
Refactor the store kit wrapper code in IAP
