+++ packages/in_app_purchase/build.yaml
@@ -4,5 +4,5 @@ targets:
       json_serializable:
         options:
           any_map: true
-          create_to_json: false
+          create_to_json: true
           nullable: false
