+++ packages/in_app_purchase/lib/src/store_kit_wrappers/sk_payment_queue_wrapper.dart
@@ -8,7 +8,9 @@ import 'package:flutter/foundation.dart';
 import 'package:in_app_purchase/src/channel.dart';
 import 'package:json_annotation/json_annotation.dart';
 import 'package:flutter/services.dart';
-import './sk_product_wrapper.dart';
+import 'sk_payment_transaction_wrappers.dart';
+import 'sk_product_wrapper.dart';
+import 'sk_download_wrapper.dart';
 
 part 'sk_payment_queue_wrapper.g.dart';
 
@@ -177,274 +179,6 @@ class SKPaymentQueueWrapper {
   }
 }
 
-/// This class is Dart wrapper around [SKTransactionObserver](https://developer.apple.com/documentation/storekit/skpaymenttransactionobserver?language=objc).
-///
-/// Must be subclassed. Must be instantiated and added to the [SKPaymentQueueWrapper] via [SKPaymentQueueWrapper.setTransactionObserver] at the app launch.
-abstract class SKTransactionObserverWrapper {
-  /// Triggered when any transactions are updated.
-  void updatedTransactions({List<SKPaymentTransactionWrapper> transactions});
-
-  /// Triggered when any transactions are removed from the payment queue.
-  void removedTransactions({List<SKPaymentTransactionWrapper> transactions});
-
-  /// Triggered when there is an error while restoring transactions.
-  void restoreCompletedTransactions({Error error});
-
-  /// Triggered when payment queue has finished sending restored transactions.
-  void paymentQueueRestoreCompletedTransactionsFinished();
-
-  /// Triggered when any download objects are updated.
-  void updatedDownloads({List<SKDownloadWrapper> downloads});
-
-  /// Triggered when a user initiated an in-app purchase from App Store.
-  ///
-  /// Return `true` to continue the transaction in your app. If you have multiple [SKTransactionObserverWrapper]s, the transaction
-  /// will continue if one [SKTransactionObserverWrapper] has [shouldAddStorePayment] returning `true`.
-  /// Return `false` to defer or cancel the transaction. For example, you may need to defer a transaction if the user is in the middle of onboarding.
-  /// You can also continue the transaction later by calling
-  /// [addPayment] with the [SKPaymentWrapper] object you get from this method.
-  bool shouldAddStorePayment(
-      {SKPaymentWrapper payment, SKProductWrapper product});
-}
-
-/// Dart wrapper around StoreKit's
-/// [SKPaymentTransactionState](https://developer.apple.com/documentation/storekit/skpaymenttransactionstate?language=objc).
-///
-/// Presents the state of a transaction. Used for handling a transaction based on different state.
-enum SKPaymentTransactionStateWrapper {
-  /// Indicates the transaction is being processed in App Store.
-  ///
-  /// You should update your UI to indicate the process and waiting for the transaction to update to the next state.
-  /// Never complete a transaction that is in purchasing state.
-  @JsonValue(0)
-  purchasing,
-
-  /// The payment is processed. You should provide the user the content they purchased.
-  @JsonValue(1)
-  purchased,
-
-  /// The transaction failed. Check the [SKPaymentTransactionWrapper.error] property from [SKPaymentTransactionWrapper] for details.
-  @JsonValue(2)
-  failed,
-
-  /// This transaction restores the content previously purchased by the user. The previous transaction information can be
-  /// obtained in [SKPaymentTransactionWrapper.originalTransaction] from [SKPaymentTransactionWrapper].
-  @JsonValue(3)
-  restored,
-
-  /// The transaction is in the queue but pending external action. Wait for another callback to get the final state.
-  ///
-  /// You should update your UI to indicate the process and waiting for the transaction to update to the next state.
-  @JsonValue(4)
-  deferred,
-}
-
-/// Dart wrapper around StoreKit's [SKPaymentTransaction](https://developer.apple.com/documentation/storekit/skpaymenttransaction?language=objc).
-///
-/// Created when a payment is added to the [SKPaymentQueueWrapper]. Transactions are delivered to your app when a payment is finished processing.
-/// Completed transactions provide a receipt and a transaction identifier that the app can use to save a permanent record of the processed payment.
-@JsonSerializable(nullable: true)
-class SKPaymentTransactionWrapper {
-  SKPaymentTransactionWrapper({
-    @required this.payment,
-    @required this.transactionState,
-    @required this.originalTransaction,
-    @required this.transactionTimeStamp,
-    @required this.transactionIdentifier,
-    @required this.downloads,
-    @required this.error,
-  });
-
-  /// Constructs an instance of this from a key value map of data.
-  ///
-  /// The map needs to have named string keys with values matching the names and
-  /// types of all of the members on this class.
-  /// The `map` parameter must not be null.
-  @visibleForTesting
-  factory SKPaymentTransactionWrapper.fromJson(Map map) {
-    if (map == null) {
-      return null;
-    }
-    return _$SKPaymentTransactionWrapperFromJson(map);
-  }
-
-  /// Current transaction state.
-  final SKPaymentTransactionStateWrapper transactionState;
-
-  /// The payment that is created and added to the payment queue which generated this transaction.
-  final SKPaymentWrapper payment;
-
-  /// The original Transaction, only available if the [transactionState] is [SKPaymentTransactionStateWrapper.restored].
-  ///
-  /// When the [transactionState] is [SKPaymentTransactionStateWrapper.restored], the current transaction object holds a new
-  /// [transactionIdentifier].
-  final SKPaymentTransactionWrapper originalTransaction;
-
-  /// The timestamp of the transaction.
-  ///
-  /// Milliseconds since epoch.
-  /// It is only defined when the [transactionState] is [SKPaymentTransactionStateWrapper.purchased] or [SKPaymentTransactionStateWrapper.restored].
-  final double transactionTimeStamp;
-
-  /// The unique string identifer of the transaction.
-  ///
-  /// It is only defined when the [transactionState] is [SKPaymentTransactionStateWrapper.purchased] or [SKPaymentTransactionStateWrapper.restored].
-  /// You may wish to record this string as part of an audit trail for App Store purchases.
-  /// The value of this string corresponds to the same property in the receipt.
-  final String transactionIdentifier;
-
-  /// An array of the [SKDownloadWrapper] object of this transaction.
-  ///
-  /// Only available if the transaction contains downloadable contents.
-  ///
-  /// It is only defined when the [transactionState] is [SKPaymentTransactionStateWrapper.purchased].
-  /// Must be used to download the transaction's content before the transaction is finished.
-  final List<SKDownloadWrapper> downloads;
-
-  /// The error object, only available if the [transactionState] is [SKPaymentTransactionStateWrapper.failed].
-  final SKError error;
-
-  @override
-  bool operator ==(Object other) {
-    if (identical(other, this)) {
-      return true;
-    }
-    if (other.runtimeType != runtimeType) {
-      return false;
-    }
-    final SKPaymentTransactionWrapper typedOther = other;
-    return typedOther.payment == payment &&
-        typedOther.transactionState == transactionState &&
-        typedOther.originalTransaction == originalTransaction &&
-        typedOther.transactionTimeStamp == transactionTimeStamp &&
-        typedOther.transactionIdentifier == transactionIdentifier &&
-        DeepCollectionEquality().equals(typedOther.downloads, downloads) &&
-        typedOther.error == error;
-  }
-}
-
-/// Dart wrapper around StoreKit's [SKDownloadState](https://developer.apple.com/documentation/storekit/skdownloadstate?language=objc).
-///
-/// The state a download operation that can be in.
-enum SKDownloadState {
-  /// Indicates that downloadable content is waiting to start.
-  @JsonValue(0)
-  waiting,
-
-  /// The downloadable content is currently being downloaded
-  @JsonValue(1)
-  active,
-
-  /// The app paused the download.
-  @JsonValue(2)
-  pause,
-
-  /// The content is successfully downloaded.
-  @JsonValue(3)
-  finished,
-
-  /// Indicates that some error occurred while the content was being downloaded.
-  @JsonValue(4)
-  failed,
-
-  /// The app canceled the download.
-  @JsonValue(5)
-  cancelled,
-}
-
-/// Dart wrapper around StoreKit's [SKDownload](https://developer.apple.com/documentation/storekit/skdownload?language=objc).
-///
-/// When a product is created in the App Store Connect, one or more download contents can be associated with it.
-/// When the product is purchased, a List of [SKDownloadWrapper] object will be present in an [SKPaymentTransactionWrapper] object.
-/// To download the content, add the [SKDownloadWrapper] objects to the payment queue and wait for the content to be downloaded.
-/// You can also read the [contentURL] to get the URL of the downloaded content after the download completes.
-/// Note that all downloaded files must be processed before the completion of the [SKPaymentTransactionWrapper]([SKPaymentQueueWrapper.finishTransaction] is called).
-/// After the transaction is complete, any [SKDownloadWrapper] object in the transaction will not be able to be added to the payment queue
-/// and the [contentURL ]of the [SKDownloadWrapper] object will be invalid.
-@JsonSerializable(nullable: true)
-class SKDownloadWrapper {
-  SKDownloadWrapper({
-    @required this.contentIdentifier,
-    @required this.state,
-    @required this.contentLength,
-    @required this.contentURL,
-    @required this.contentVersion,
-    @required this.transactionID,
-    @required this.progress,
-    @required this.timeRemaining,
-    @required this.downloadTimeUnknown,
-    @required this.error,
-  });
-
-  /// Constructs an instance of this from a key value map of data.
-  ///
-  /// The map needs to have named string keys with values matching the names and
-  /// types of all of the members on this class.
-  /// The `map` parameter must not be null.
-  @visibleForTesting
-  factory SKDownloadWrapper.fromJson(Map map) {
-    assert(map != null);
-    return _$SKDownloadWrapperFromJson(map);
-  }
-
-  /// Identifies the downloadable content.
-  ///
-  /// It is specified in the App Store Connect when the downloadable content is created.
-  final String contentIdentifier;
-
-  /// The current download state.
-  ///
-  /// When the state changes, one of the [SKTransactionObserverWrapper] subclasses' observing methods should be triggered.
-  /// The developer should properly handle the downloadable content based on the state.
-  final SKDownloadState state;
-
-  /// Length of the content in bytes.
-  final int contentLength;
-
-  /// The URL string of the content.
-  final String contentURL;
-
-  /// Version of the content formatted as a series of dot-separated integers.
-  final String contentVersion;
-
-  /// The transaction ID of the transaction that is associated with the downloadable content.
-  final String transactionID;
-
-  /// The download progress, between 0.0 to 1.0.
-  final double progress;
-
-  /// The estimated time remaining for the download; if no good estimate is able to be made,
-  /// [downloadTimeUnknown] will be set to true.
-  final double timeRemaining;
-
-  /// true if [timeRemaining] cannot be estimated.
-  final bool downloadTimeUnknown;
-
-  /// The error that prevented the downloading; only available if the [transactionState] is [SKPaymentTransactionStateWrapper.failed].
-  final SKError error;
-
-  @override
-  bool operator ==(Object other) {
-    if (identical(other, this)) {
-      return true;
-    }
-    if (other.runtimeType != runtimeType) {
-      return false;
-    }
-    final SKDownloadWrapper typedOther = other;
-    return typedOther.contentIdentifier == contentIdentifier &&
-        typedOther.state == state &&
-        typedOther.contentLength == contentLength &&
-        typedOther.contentURL == contentURL &&
-        typedOther.contentVersion == contentVersion &&
-        typedOther.transactionID == transactionID &&
-        typedOther.progress == progress &&
-        typedOther.timeRemaining == timeRemaining &&
-        typedOther.downloadTimeUnknown == downloadTimeUnknown &&
-        typedOther.error == error;
-  }
-}
-
 /// Dart wrapper around StoreKit's [NSError](https://developer.apple.com/documentation/foundation/nserror?language=objc).
 @JsonSerializable(nullable: true)
 class SKError {
@@ -456,7 +190,6 @@ class SKError {
   /// The map needs to have named string keys with values matching the names and
   /// types of all of the members on this class.
   /// The `map` parameter must not be null.
-  @visibleForTesting
   factory SKError.fromJson(Map map) {
     assert(map != null);
     return _$SKErrorFromJson(map);
@@ -506,7 +239,6 @@ class SKPaymentWrapper {
   /// The map needs to have named string keys with values matching the names and
   /// types of all of the members on this class.
   /// The `map` parameter must not be null.
-  @visibleForTesting
   factory SKPaymentWrapper.fromJson(Map map) {
     assert(map != null);
     return _$SKPaymentWrapperFromJson(map);
@@ -564,4 +296,7 @@ class SKPaymentWrapper {
         typedOther.simulatesAskToBuyInSandbox == simulatesAskToBuyInSandbox &&
         typedOther.requestData == requestData;
   }
+
+  @override
+  String toString() => _$SKPaymentWrapperToJson(this).toString();
 }
