+++ packages/in_app_purchase/lib/src/store_kit_wrappers/sk_product_wrapper.dart
@@ -325,7 +325,7 @@ class SKPriceLocaleWrapper {
   /// The `map` parameter must not be null.
   factory SKPriceLocaleWrapper.fromJson(Map map) {
     assert(map != null, 'Map must not be null.');
-    return _$PriceLocaleWrapperFromJson(map);
+    return _$SKPriceLocaleWrapperFromJson(map);
   }
 
   ///The currency symbol for the locale, e.g. $ for US locale.
