+++ packages/in_app_purchase/lib/store_kit_wrappers.dart
@@ -2,7 +2,9 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+export 'src/store_kit_wrappers/sk_download_wrapper.dart';
 export 'src/store_kit_wrappers/sk_payment_queue_wrapper.dart';
+export 'src/store_kit_wrappers/sk_payment_transaction_wrappers.dart';
 export 'src/store_kit_wrappers/sk_product_wrapper.dart';
+export 'src/store_kit_wrappers/sk_receipt_manager.dart';
 export 'src/store_kit_wrappers/sk_request_maker.dart';
-export 'src/store_kit_wrappers//sk_receipt_manager.dart';
